# planning-poker/internal/infra/boundaries/http/getroom.go
package http

import (
	"errors"
	"net/http"
	"planning-poker/internal/domain"

	"github.com/bruno303/go-toolkit/pkg/log"
	"github.com/gorilla/mux"
)

type (
	GetRoomResponse struct {
		RoomID string `json:"roomId"`
	}
	GetRoomAPI struct {
		hub    domain.Hub
		logger log.Logger
	}
)

var _ API = (*GetRoomAPI)(nil)

// @Summary Get room information
// @Description Returns information about a specific room
// @Tags rooms
// @Produce json
// @Param roomID path string true "Room ID"
// @Success 200 {object} GetRoomResponse
// @Failure 404 {object} ErrorResponse
// @Router /planning/room/{roomID} [get]
func NewGetRoomAPI(hub domain.Hub) GetRoomAPI {
	return GetRoomAPI{
		hub:    hub,
		logger: log.NewLogger("getroomapi"),
	}
}

func (g GetRoomAPI) Endpoint() string {
	return "/planning/room/{roomID}"
}

func (g GetRoomAPI) Methods() []string {
	return []string{"GET"}
}

func (g GetRoomAPI) Handle() http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := r.Context()
		roomID := mux.Vars(r)["roomID"]
		if roomID == "" {
			SendJsonErrorMsg(w, http.StatusBadRequest, "Room ID is required")
			return
		}

		_, err := g.hub.LoadRoom(ctx, roomID)
		if errors.Is(err, domain.ErrRoomNotFound) {
			SendJsonErrorMsg(w, http.StatusNotFound, "Room not found")
			return
		}
		if err != nil {
			g.logger.Error(ctx, "Failed to load room", err)
			SendJsonErrorMsg(w, http.StatusInternalServerError, "Failed to load room")
			return
		}

		SendJsonResponse(w, http.StatusOK, GetRoomResponse{RoomID: roomID})
	})
}